requirements.txt
bleak
Pillow
pymupdf
